Update version_control_pattern.md
Felt that the changing your text editor bit was a bit confusing and tried to take it out of the git-flow by formatting it differently (not sure if I accomplished this enough though). Also added a link to another article that is related to the first link in this guide but is more accessible.